common text handler for comment and reply
